refactor: rename and relocate files and add partial of GET API for get user data in profile page
